+++ example/android/app/src/main/AndroidManifest.xml
@@ -24,22 +24,14 @@
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
-            <!-- Deep Link Intent Filter for MRTD Validation -->
-            <intent-filter android:autoVerify="true">
-                <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="mrtd" 
-                      android:host="validate" />
-            </intent-filter>
-            <!-- Additional intent filter for HTTP/HTTPS fallback -->
+            <!-- App Link for https://passport-issuer.staging.yivi.app/start-app -->
             <intent-filter android:autoVerify="true">
                 <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
                 <data android:scheme="https"
-                      android:host="mrtd.app"
-                      android:pathPrefix="/validate" />
+                      android:host="passport-issuer.staging.yivi.app"
+                      android:pathPrefix="/start-app" />
             </intent-filter>
         </activity>
         <!-- Don't delete the meta-data below.
@@ -48,6 +40,7 @@
             android:name="flutterEmbedding"
             android:value="2" />
     </application>
+    
     <!-- Required to query activities that can process text, see:
          https://developer.android.com/training/package-visibility and
          https://developer.android.com/reference/android/content/Intent#ACTION_PROCESS_TEXT.
