+++ example/android/app/src/main/kotlin/app/yivi/vcmrtd/MainActivity.kt
